Get working on 32-bit

+++ codex-rs/linux-sandbox/src/landlock.rs
@@ -209,8 +209,8 @@ fn install_network_seccomp_filter_on_current_thread(
                 libc::AF_UNIX as u64,
             )?])?;
 
-            rules.insert(libc::SYS_socket, vec![unix_only_rule.clone()]);
-            rules.insert(libc::SYS_socketpair, vec![unix_only_rule]);
+            rules.insert(i64::from(libc::SYS_socket), vec![unix_only_rule.clone()]);
+            rules.insert(i64::from(libc::SYS_socketpair), vec![unix_only_rule]);
         }
         NetworkSeccompMode::ProxyRouted => {
             // In proxy-routed mode we allow IP sockets in the isolated
@@ -238,8 +238,8 @@ fn install_network_seccomp_filter_on_current_thread(
                 SeccompCmpOp::Eq,
                 libc::AF_UNIX as u64,
             )?])?;
-            rules.insert(libc::SYS_socket, vec![deny_non_ip_socket]);
-            rules.insert(libc::SYS_socketpair, vec![deny_unix_socketpair]);
+            rules.insert(i64::from(libc::SYS_socket), vec![deny_non_ip_socket]);
+            rules.insert(i64::from(libc::SYS_socketpair), vec![deny_unix_socketpair]);
         }
     }
